made GameOnScreen component
