+++ code/src/components/StartScreen.js
@@ -33,8 +33,6 @@ const StartScreen = () => {
                 <NameInput
                     type="text"
                     id="name_input"
-                    pattern=".{4,25}"
-                    required
                     value={inputValue}
                     onChange={event => setInputValue(event.target.value)}
                 />
